+++ src/commands/hardis/org/configure/monitoring.ts
@@ -85,7 +85,7 @@ export default class OrgConfigureMonitoring extends SfdxCommand {
         throw new SfdxError('Your git repository name must contain the expression "monitoring"');
       }
     }
-    const preRequisitesUrl = "https://sfdx-hardis.cloudity.com/salesforce-monitoring-home/#configuration";
+    const preRequisitesUrl = "https://sfdx-hardis.cloudity.com/salesforce-monitoring-config-home/#instructions";
     uxLog(this, c.yellow("Monitoring pre-requisites documentation: " + c.bold(preRequisitesUrl)));
     const confirmPreRequisites = await prompts({
       type: "select",
@@ -197,7 +197,7 @@ export default class OrgConfigureMonitoring extends SfdxCommand {
     }
     const branch = await getCurrentGitBranch();
     uxLog(this, c.greenBright(`Now you must schedule monitoring to run the job automatically every night on branch ${c.bold(branch)}:)`));
-    const scheduleMonitoringUrl = "https://sfdx-hardis.cloudity.com/salesforce-monitoring-home/#configuration";
+    const scheduleMonitoringUrl = "https://sfdx-hardis.cloudity.com/salesforce-monitoring-config-home/#instructions";
     const msg = "Please follow the instructions to schedule sfdx-hardis monitoring on your Git server: " + c.bold(scheduleMonitoringUrl);
     uxLog(this, c.yellow(msg));
     await open(scheduleMonitoringUrl, { wait: true });